fix: refactor removeBucketIamMember.js
