+++ samples/iam.js
@@ -89,61 +89,6 @@ async function addBucketIamMember(bucketName, roleName, members) {
   // [END storage_add_bucket_iam_member]
 }
 
-async function removeBucketIamMember(bucketName, roleName, members) {
-  // [START storage_remove_bucket_iam_member]
-  // Imports the Google Cloud client library
-  const {Storage} = require('@google-cloud/storage');
-
-  /**
-   * TODO(developer): Uncomment the following lines before running the sample.
-   */
-  // const bucketName = 'Name of a bucket, e.g. my-bucket';
-  // const roleName = 'Role to grant, e.g. roles/storage.objectViewer';
-  // const members = [
-  //   'user:jdoe@example.com',    // Example members to grant
-  //   'group:admins@example.com', // the new role to
-  // ];
-
-  // Creates a client
-  const storage = new Storage();
-
-  // Get a reference to a Google Cloud Storage bucket
-  const bucket = storage.bucket(bucketName);
-
-  // Gets and updates the bucket's IAM policy
-  const [policy] = await bucket.iam.getPolicy();
-
-  // Finds and updates the appropriate role-member group
-  const index = policy.bindings.findIndex(role => role.role === roleName);
-  const role = policy.bindings[index];
-  if (role) {
-    role.members = role.members.filter(
-      member => members.indexOf(member) === -1
-    );
-
-    // Updates the policy object with the new (or empty) role-member group
-    if (role.members.length === 0) {
-      policy.bindings.splice(index, 1);
-    } else {
-      policy.bindings.index = role;
-    }
-
-    // Updates the bucket's IAM policy
-    await bucket.iam.setPolicy(policy);
-  } else {
-    // No matching role-member group(s) were found
-    throw new Error('No matching role-member group(s) found.');
-  }
-
-  console.log(
-    `Removed the following member(s) with role ${roleName} from ${bucketName}:`
-  );
-  members.forEach(member => {
-    console.log(`  ${member}`);
-  });
-  // [END storage_remove_bucket_iam_member]
-}
-
 require(`yargs`)
   .demand(1)
   .array('members')
@@ -158,12 +103,6 @@ require(`yargs`)
     `Adds one or more IAM member-role groups to a Google Cloud Storage bucket.`,
     {},
     opts => addBucketIamMember(opts.bucketName, opts.roleName, opts.members)
-  )
-  .command(
-    `remove-members <bucketName> <roleName> [members..]`,
-    `Removes one or more IAM member-role groups from a Google Cloud Storage bucket.`,
-    {},
-    opts => removeBucketIamMember(opts.bucketName, opts.roleName, opts.members)
   )
   .example(`node $0 view-members "my-bucket"`)
   .example(
